support for 64 bit ints
required for storing time

+++ schema/mysql.sql
@@ -11,7 +11,7 @@ CREATE TABLE `json` (
   `path9` varchar(32) DEFAULT NULL,
   `type` varchar(15) NOT NULL,
   `bool_value` tinyint(1) DEFAULT NULL,
-  `int_value` int(11) DEFAULT NULL,
+  `int_value` bigint(20) DEFAULT NULL,
   `varchar_value` varchar(255) DEFAULT NULL,
   `text_value` text DEFAULT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
